layout/forms/test/test_bug478219.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=478219
-->
<head>
  <title>Test for Bug 478219</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style type="text/css">

  </style>
</head>

<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=478219">Mozilla Bug 478219</a>
<input style="overflow: scroll;-moz-binding:url(#xbl);" type="image"/>
<pre id="test">
<script class="testbody" type="text/javascript">
  var win = window.open('', '', 'width=100,height=100');
  win.close();
  is(win.closed, true, "Window should have been opened");
  is(win.opener, window, "Opened window should have as opener this window");
</script>
</pre>

<bindings xmlns="http://www.mozilla.org/xbl">
<binding id="xbl" inheritstyle="false">
<content>

<div xmlns="http://www.w3.org/1999/xhtml" contenteditable="true">
<style>style {-moz-binding:url(#xbl);</style>
</div>
<textarea xmlns="http://www.w3.org/1999/xhtml"></textarea>
</content>
</binding>
</bindings>
</body>
</html>
